:pencil2: Organized Header and Brand components into Layout folder
