+++ app/layout.js
@@ -1,7 +1,7 @@
 import "./globals.css"
 import { Inter } from "next/font/google"
 
-import Header from "@/components/header"
+import Header from "@/components/Layout/Header"
 
 const inter = Inter({ subsets: ["latin"] })
